# openapi.yaml
...
openapi: "3.0.3"
info:
  title: "identuum-idp-oss API"
  version: "identuum-idp-oss-api-docgen 0.2.0-p3"
  description: "Machine-generated OpenAPI 3 specification for the identuum-idp-oss API surface. Generated by tools/api-docgen from // docgen: source annotations. Response bodies without a known schema record are described but not field-modeled; see the referenced Go symbol in identuum-idp-oss for the authoritative shape."
paths:
...
  "/api/v1/auth/idp/{id}/callback":
    get:
      operationId: "identuum-idp-oss.oidc-login.oidc_callback"
      summary: "Complete upstream single-provider OIDC login: consume the one-time state, exchange the code for tokens, strictly validate the ID token against the provider JWKS, JIT-provision or match the local user, mint a local session, and 302-redirect to the stored return URL with the session cookie set."
      tags:
        - "oidc-login"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security: []
      responses:
        "200":
          description: "Complete upstream single-provider OIDC login: consume the one-time state, exchange the code for tokens, strictly validate the ID token against the provider JWKS, JIT-provision or match the local user, mint a local session, and 302-redirect to the stored return URL with the session cookie set."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
...
  "/api/v1/oidc/logout":
    get:
      operationId: "identuum-idp-oss.oidc.end_session"
      summary: "OIDC RP-initiated logout / end_session endpoint (clears the browser cookie + revokes the cookie-resolved session; honours post_logout_redirect_uri when allowed by the registered client)."
      tags:
        - "oidc"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "OIDC RP-initiated logout / end_session endpoint (clears the browser cookie + revokes the cookie-resolved session; honours post_logout_redirect_uri when allowed by the registered client)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: "identuum_session"
      description: "Browser session cookie backing a cookie-authenticated Principal (site_admin, org_admin, or any authenticated user, depending on the operation's actual role requirement — see its summary)."
    oauthClientAuth:
      type: http
      scheme: basic
      description: "OAuth 2.1 client authentication (client credentials / RFC 7592 management token) at client-management and token endpoints."
...